refactor(meeting): split manager.rs test_support and events into sub-modules (#708)
